feat: RabbitConfig 수정
