+++ src/main/java/com/example/teamrocket/config/rabbit/RabbitConfig.java
@@ -1,0 +1,82 @@
+package com.example.teamrocket.config.rabbit;
+
+import com.fasterxml.jackson.databind.Module;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import org.springframework.amqp.core.Binding;
+import org.springframework.amqp.core.BindingBuilder;
+import org.springframework.amqp.core.Queue;
+import org.springframework.amqp.core.TopicExchange;
+import org.springframework.amqp.rabbit.annotation.EnableRabbit;
+import org.springframework.amqp.rabbit.connection.CachingConnectionFactory;
+import org.springframework.amqp.rabbit.connection.ConnectionFactory;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableRabbit
+public class RabbitConfig {
+
+    private static final String CHAT_QUEUE_NAME = "chat.queue";
+    private static final String CHAT_EXCHANGE_NAME = "chat.exchange";
+    private static final String ROUTING_KEY = "room.*";
+
+    @Value("${spring.rabbitmq.username}")
+    private String adminName;
+    @Value("${spring.rabbitmq.password}")
+    private String adminPassword;
+    @Value("${spring.rabbitmq.host}")
+    private String host;
+
+    //Queue 등록
+    @Bean
+    public Queue queue(){ return new Queue(CHAT_QUEUE_NAME, true); }
+
+    //Exchange 등록
+    @Bean
+    public TopicExchange exchange(){ return new TopicExchange(CHAT_EXCHANGE_NAME); }
+
+    //Exchange와 Queue 바인딩
+    @Bean
+    public Binding binding(Queue queue, TopicExchange exchange) {
+        return BindingBuilder.bind(queue).to(exchange).with(ROUTING_KEY);
+    }
+
+    /* messageConverter를 커스터마이징 하기 위해 Bean 새로 등록 */
+    @Bean
+    public RabbitTemplate rabbitTemplate(){
+        RabbitTemplate rabbitTemplate = new RabbitTemplate(connectionFactory());
+        rabbitTemplate.setMessageConverter(jsonMessageConverter());
+        return rabbitTemplate;
+    }
+
+    //Spring에서 자동생성해주는 ConnectionFactory는 SimpleConnectionFactory인가? 그건데
+    //여기서 사용하는 건 CachingConnectionFacotry라 새로 등록해줌
+    @Bean
+    public ConnectionFactory connectionFactory(){
+        CachingConnectionFactory factory = new CachingConnectionFactory();
+        factory.setHost(host);
+        factory.setUsername(adminName);
+        factory.setPassword(adminPassword);
+        return factory;
+    }
+
+    @Bean
+    public Jackson2JsonMessageConverter jsonMessageConverter(){
+        //LocalDateTime serializable을 위해
+        ObjectMapper objectMapper = new ObjectMapper();
+        objectMapper.configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, true);
+        objectMapper.registerModule(dateTimeModule());
+
+        return new Jackson2JsonMessageConverter(objectMapper);
+    }
+
+    @Bean
+    public Module dateTimeModule(){
+        return new JavaTimeModule();
+    }
+}
